startup: Remove clean_pyc; log a warning if ./src/ is detected.
Summary:
This caused a race condition that would break startup when multiple
parallel sync engines were started at the same time.

clean_pyc() should be a no-op now anyway, since our requirements.txt no
longer has any `-e git+https://` lines in it.

Also:
- Removed `sys.dont_write_bytecode = True`.  It doesn't really make
  sense to do it here, since we've already started importing things and
  writing .pyc files at this point.  If we actually want to avoid
  writing .pyc files at all (which I doubt), we should set
  PYTHONDONTWRITEBYTECODE=1 in the environment before starting Python.

- Moved `find . -name \*.pyc -delete` to earlier in the setup process
  so that old pyc files are removed before we start installing Python
  stuff, rather than afterwards.

Test Plan:
Run bin/inbox-start once under Vagrant to make sure it
doesn't crash.

Reviewers: charles

Reviewed By: charles

Maniphest Tasks: T463

Differential Revision: https://review.inboxapp.com/D479

+++ setup.sh
@@ -34,6 +34,11 @@ color '36;1' "
      https://www.github.com/inboxapp/inbox
 "
 
+if ! [ -e ./setup.py ] || ! [ -e ./setup.sh ] ; then
+    color '31;1' "Error: setup.sh should be run from the inbox repo" >&2
+    exit 1
+fi
+
 color '35;1' 'Updating packages...'
 apt-get update
 apt-get -y install python-software-properties
@@ -65,6 +70,9 @@ apt-get -y install git \
                    tmux \
                    curl \
                    tnef \
+
+color '35;1' 'Removing .pyc files...'   # they might be stale
+find . -name \*.pyc -delete
 
 color '35;1' 'Ensuring setuptools and pip versions...'
 # If python-setuptools is actually the old 'distribute' fork of setuptools,
@@ -158,9 +166,6 @@ if $configure_db; then
         alembic upgrade head
     fi
 fi
-
-color '35;1' 'Removing .pyc files...'
-find . -name \*.pyc -delete
 
 color '35;1' 'Cleaning up...'
 apt-get -y purge build-essential
